requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
anthropic>=0.18.0

# Training
accelerate>=0.24.0
peft>=0.7.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Logging & monitoring
wandb>=0.16.0
tensorboard>=2.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
